+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Changelog
 This file contains highlights of what changes on each version of the [Mustache for Yii](https://github.com/cedx/yii2-mustache) library.
 
+## Version 8.0.0
+- Added the `ViewRenderer::$enableCaching` property.
+
 ## Version 7.0.0
 - Breaking change: renamed the `yii\mustache\helpers\HTML` class to `Html`.
 - Breaking change: renamed the `yii\mustache\helpers\URL` class to `Url`.